# openapi/openapi.json
{
  "components": {
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "BAD_REQUEST",
                "message": "Request failed"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Bad request"
      },
...
      "Forbidden": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "FORBIDDEN",
                "message": "Insufficient permissions"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Authenticated principal lacks permission"
      },
      "GenericSuccess": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenericSuccessResponse"
            }
          }
        },
        "description": "Successful JSON response using the standard API envelope"
      },
      "InternalError": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "INTERNAL_ERROR",
                "message": "An unexpected error occurred"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Unexpected server error"
      },
      "NotFound": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "NOT_FOUND",
                "message": "Resource not found"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Resource or route not found"
      },
...
      "TooManyRequests": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "TOO_MANY_REQUESTS",
                "message": "Too many requests"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Rate limit exceeded"
      },
      "Unauthorized": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Authentication is missing, invalid or expired"
      },
      "ValidationError": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "VALIDATION_ERROR",
                "details": [],
                "message": "Request validation failed"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Validation failed"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "success": {
            "enum": [
              false
            ],
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "error"
        ],
        "type": "object"
      },
      "GenericData": {
        "additionalProperties": true,
        "description": "Endpoint-specific payload retained as an extensible object until its domain schema is specialized.",
        "type": "object"
      },
      "GenericSuccessResponse": {
        "additionalProperties": false,
        "example": {
          "data": {},
          "success": true
        },
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GenericData"
          },
          "message": {
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/GenericData"
          },
          "success": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Carbon tracking and sustainability management API for WeaveCarbon.",
    "title": "WeaveCarbon API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/evidence": {
      "get": {
        "operationId": "getEvidence",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "productId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /evidence",
        "tags": [
          "Evidence"
        ]
      },
      "post": {
        "operationId": "postEvidence",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /evidence",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/product/{product_id}": {
      "get": {
        "operationId": "getEvidenceProductByProductId",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /evidence/product/{product_id}",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/upload": {
      "post": {
        "operationId": "postEvidenceUpload",
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "additionalProperties": true,
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /evidence/upload",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/{id}": {
      "delete": {
        "operationId": "deleteEvidenceById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "DELETE /evidence/{id}",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/{id}/confirm": {
      "post": {
        "operationId": "postEvidenceByIdConfirm",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /evidence/{id}/confirm",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/{id}/fields": {
      "get": {
        "operationId": "getEvidenceByIdFields",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /evidence/{id}/fields",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/{id}/lock": {
      "post": {
        "operationId": "postEvidenceByIdLock",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /evidence/{id}/lock",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/{id}/rag-ingest": {
      "post": {
        "operationId": "postEvidenceByIdRagIngest",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "additionalProperties": true,
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /evidence/{id}/rag-ingest",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/{id}/status": {
      "get": {
        "operationId": "getEvidenceByIdStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /evidence/{id}/status",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/{id}/verify": {
      "post": {
        "operationId": "postEvidenceByIdVerify",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /evidence/{id}/verify",
        "tags": [
          "Evidence"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Current environment",
      "url": "/api"
    }
  ],
...
}